@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { ApiService } from '../api.service';
 import { Router } from '@angular/router';
+import { AuthService } from '../auth.service';
+import { ApiService } from '../api.service';
 
 @Component({
   selector: 'app-login',
@@ -8,34 +9,49 @@ import { Router } from '@angular/router';
   styleUrls: ['./login.component.css']
 })
 export class LoginComponent {
-  user = '';
-  pass = '';
-  otp = '';
-  step = 1 ;
+  username = '';
+  password = '';
+  error = '';
+  loading = false;
 
-  constructor(private api: ApiService, private router: Router){}
+  constructor(
+    private authService: AuthService,
+    private apiService: ApiService,
+    private router: Router
+  ) { }
 
-  onLogin(){
-    this.api.login(this.user, this.pass).subscribe({
-      next: (res)=>{
-        alert("OTP Generated: " + res);
-        this.step = 2;
-      },
-      error: () => alert("Invalid Credentials")
-    });
-  }
-  onVerify(){
-    this.api.verify(this.user, this.otp).subscribe({
-      next: (res)=>{
+  onLogin() {
+    this.error = '';
+    this.loading = true;
 
-        localStorage.setItem('user', JSON.stringify(res));
-        if(res.role === 'OWNER'){
-          this.router.navigate(['/owner']);
-        } else{
-          this.router.navigate(['/dashboard'])
+    // First set credentials for Basic Auth header
+    this.authService.login(this.username, this.password, 0, '');
+
+    // Call login API to get account details for the username
+    this.apiService.login(this.username).subscribe({
+      next: (response) => {
+        this.loading = false;
+        if (response.accountId) {
+          // Update auth service with full account details
+          this.authService.login(this.username, this.password, response.accountId, response.holderName);
+          // Store account info
+          localStorage.setItem('account', JSON.stringify({
+            id: response.accountId,
+            holderName: response.holderName,
+            balance: response.balance,
+            status: response.status
+          }));
+          this.router.navigate(['/dashboard']);
+        } else {
+          this.authService.logout();
+          this.error = response.message || 'Login failed. Please try again.';
         }
       },
-      error: () => alert("Invalid OTP")
+      error: (err) => {
+        this.loading = false;
+        this.authService.logout();
+        this.error = 'Invalid credentials. Please try again.';
+      }
     });
   }
 }
